refactor(iota-multiaddr): own the wasm stand-in in the crate
The wasm stand-in lived in `iota-types`, which meant a crate carried a
substitute for a type it doesn't own, and had to pick between the two with
a `cfg` switch on the re-export.

- split the crate into `native` and `wasm` modules and select between them
  in `lib.rs`
- move `wasm_multiaddr.rs` from `iota-types` in as the `wasm` module
- gate the native-only dependencies behind
  `cfg(not(target_arch = "wasm32"))`; the stand-in needs only serde
- re-export unconditionally from `iota-types` and move the dependency out
  of its non-wasm section

The public API still differs by target, since the parsing and anemo helpers
are native-only; `lib.rs` and the `wasm` module docs say so.
