+++ crates/iota-types/Cargo.toml
@@ -49,6 +49,7 @@ tracing.workspace = true
 
 # internal dependencies
 iota-macros.workspace = true
+iota-multiaddr.workspace = true
 iota-protocol-config.workspace = true
 iota-sdk-crypto = { workspace = true, features = ["bech32", "ed25519", "passkey", "rand", "secp256k1", "secp256r1"] }
 iota-sdk-types = { workspace = true, features = ["serde", "rand", "hash"] }
@@ -68,7 +69,6 @@ fastcrypto-tbls.workspace = true
 tonic.workspace = true
 
 # internal dependencies
-iota-multiaddr.workspace = true
 prometheus-filtered.workspace = true
 
 [dev-dependencies]
